gestion_emballage/afrikpros/requirement.txt
asgiref==3.7.2
django-browser-reload==1.12.1
sqlparse==0.4.4
typing_extensions==4.9.0
